move session handling out of the user class
